Affinity and DigitalOcean workarounds
DigitalOcean K8s doesn't support the ReadWriteMany access mode for PVCs.
This is required to share bridge/appservice registration files with
Synapse, as the bridge pods are not necessarily scheduled on the same
node as Synapse.

The workaround was to:
 - Add an override to change the access mode of the bridge PVC in
   values.yaml
 - Add a setting to configure affinity between bridges and Synapse, so
   that this access mode is actually valid

DOK8S also requires a minimum PVC size of 1Gi. Added an override for the
size of the bridge PVC, but left the default at 1Mi.